requirements.txt
aiohttp~=3.8.3
asyncio~=3.4.3
aiomysql
discord.py
nest_asyncio
mysql-connector-python~=8.0.31
PyNaCl
praw
igdb-api-v4
requests~=2.28.1
ratelimit~=2.2.1